refactor: split iframe helpers into dedicated files
